feat(bonds): add discount bonds
