+++ packages/bonds/contracts/libs/Adminable.sol
@@ -1,0 +1,25 @@
+// SPDX-License-Identifier: GPL-3.0
+pragma solidity 0.8.9;
+
+abstract contract Adminable {
+    event AdminUpdated(address indexed user, address indexed newAdmin);
+
+    address public admin;
+
+    modifier onlyAdmin() virtual {
+        require(msg.sender == admin, "UNAUTHORIZED");
+
+        _;
+    }
+
+    function setAdmin(address newAdmin) public virtual onlyAdmin {
+        _setAdmin(newAdmin);
+    }
+
+    function _setAdmin(address newAdmin) internal {
+        require(newAdmin != address(0), "Can not set admin to zero address");
+        admin = newAdmin;
+
+        emit AdminUpdated(msg.sender, newAdmin);
+    }
+}
