--- packages/bonds/deployments/bsc/.chainId
@@ -1,1 +1,0 @@
-56
