--- packages/bonds/deployments/bsc/.extraMeta/AccidentHandler20220715V3.json
@@ -1,4 +1,0 @@
-{
-  "class": "AccidentHandler20220715V3",
-  "instance": "AccidentHandler20220715V3"
-}
