+++ packages/bonds/deployments/bsctest/.extraMeta/BondFactory.json
@@ -1,0 +1,4 @@
+{
+  "class": "BondFactory",
+  "instance": "BondFactory"
+}
