+++ packages/bonds/deployments/bsctest/.extraMeta/DiscountBond.json
@@ -1,0 +1,4 @@
+{
+  "class": "DiscountBond",
+  "instance": "DiscountBond"
+}
