+++ packages/bonds/deployments/bsctest/.extraMeta/DuetTransparentUpgradeableProxy.json
@@ -1,0 +1,4 @@
+{
+  "class": "DuetTransparentUpgradeableProxy",
+  "instance": "DuetTransparentUpgradeableProxy"
+}
