+++ packages/bonds/package.json
@@ -12,8 +12,8 @@
     "compile": "npx hardhat compile",
     "compile:watch": "npx hardhat watch compile",
     "clean": "rm -rf artifacts && rm -rf cache && rm -rf coverage && rm -rf typechain",
-    "deploy:bsc": "npx hardhat deploy --network bsc && ts-node -T ./scripts/clean-solc-inputs.ts",
-    "deploy:bsctest": "npx hardhat deploy --network bsctest && ts-node -T ./scripts/clean-solc-inputs.ts",
+    "deploy:bsc": "pnpm run prettier && npx hardhat deploy --network bsc && ts-node -T ./scripts/clean-solc-inputs.ts",
+    "deploy:bsctest": "pnpm run prettier && npx hardhat deploy --network bsctest && ts-node -T ./scripts/clean-solc-inputs.ts",
     "deploy:local": "npx hardhat deploy --network local",
     "verify:bsc": "npx hardhat --network bsc etherscan-verify && npx hardhat --network bsc verify:duet",
     "verify:bsctest": "npx hardhat --network bsctest etherscan-verify && npx hardhat --network bsctest verify:duet",
