--- packages/bonds/scripts/compensasion.ts
@@ -1,7 +1,0 @@
-export const tokens = {
-  BUSD: '0xe9e7cea3dedca5984780bafc599bd69add087d56',
-  'dWTI-BUSD': '0x3ebd1b7a27dcc113639fce3a063d6083ef0ef56d',
-  'dXAU-BUSD': '0x95ca57ff396864c25520bc97deaae978daaf73f3',
-  'dTMC-BUSD': '0xf048897b35963aaed1a241512d26540c7ec42a60',
-  'DUET-CAKE': '0xbdf0aa1d1985caa357a6ac6661d838da8691c569',
-}
